requirements.txt
python-dotenv>=1.0.0
pandas>=1.5.0
numpy>=1.21.0
yfinance>=0.2.0
pytz>=2023.3
requests>=2.28.0
alpaca-py>=0.18.0
scikit-learn>=1.0.0
